+++ lib/widget/part_detail.dart
@@ -77,7 +77,7 @@ class _PartDisplayState extends RefreshableState<PartDetailWidget> {
 
   @override
   Future<void> onBuild(BuildContext context) async {
-    refresh();
+    refresh(context);
 
     setState(() {
 
@@ -85,8 +85,15 @@ class _PartDisplayState extends RefreshableState<PartDetailWidget> {
   }
 
   @override
-  Future<void> request() async {
-    await part.reload();
+  Future<void> request(BuildContext context) async {
+
+    final bool result = await part.reload();
+
+    if (!result || part.pk == -1) {
+      // Part could not be loaded, for some reason
+      Navigator.of(context).pop();
+    }
+
     await part.getTestTemplates();
   }
 
@@ -94,7 +101,7 @@ class _PartDisplayState extends RefreshableState<PartDetailWidget> {
 
     if (InvenTreeAPI().checkPermission("part", "view")) {
       await part.update(values: {"starred": "${!part.starred}"});
-      refresh();
+      refresh(context);
     }
   }
 
@@ -104,7 +111,7 @@ class _PartDisplayState extends RefreshableState<PartDetailWidget> {
       context,
       L10().editPart,
       onSuccess: (data) async {
-        refresh();
+        refresh(context);
         showSnackIcon(L10().partEdited, success: true);
       }
     );
@@ -130,7 +137,7 @@ class _PartDisplayState extends RefreshableState<PartDetailWidget> {
                   builder: (context) => PartImageWidget(part)
                 )
               ).then((value) {
-                refresh();
+                refresh(context);
               });
             }),
         ),