+++ lib/widget/stock_detail.dart
@@ -96,13 +96,20 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
 
     // Load part data if not already loaded
     if (part == null) {
-      refresh();
+      refresh(context);
     }
   }
 
   @override
-  Future<void> request() async {
-    await item.reload();
+  Future<void> request(BuildContext context) async {
+
+    final bool result = await item.reload();
+
+    // Could not load this stock item for some reason
+    // Perhaps it has been depleted?
+    if (!result || item.pk == -1) {
+      Navigator.of(context).pop();
+    }
 
     // Request part information
     part = await InvenTreePart().get(item.partId) as InvenTreePart?;
@@ -147,6 +154,27 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
         });
       }
     });
+  }
+
+  /// Delete the stock item from the database
+  Future<void> _deleteItem(BuildContext context) async {
+
+    confirmationDialog(
+      L10().stockItemDelete,
+      L10().stockItemDeleteConfirm,
+      icon: FontAwesomeIcons.trashAlt,
+      onAccept: () async {
+        final bool result = await item.delete();
+        
+        if (result) {
+          Navigator.of(context).pop();
+          showSnackIcon(L10().stockItemDeleteSuccess, success: true);
+        } else {
+          showSnackIcon(L10().stockItemDeleteFailure, success: false);
+        }
+      },
+    );
+
   }
 
   /// Opens a popup dialog allowing user to select a label for printing
@@ -244,7 +272,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
       L10().editItem,
       fields: fields,
       onSuccess: (data) async {
-        refresh();
+        refresh(context);
         showSnackIcon(L10().stockItemUpdated, success: true);
       }
     );
@@ -261,7 +289,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
 
     _stockUpdateMessage(result);
 
-    refresh();
+    refresh(context);
   }
 
   Future <void> _addStockDialog() async {
@@ -293,7 +321,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
         icon: FontAwesomeIcons.plusCircle,
         onSuccess: (data) async {
           _stockUpdateMessage(true);
-          refresh();
+          refresh(context);
         }
       );
 
@@ -340,7 +368,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
 
     _stockUpdateMessage(result);
 
-    refresh();
+    refresh(context);
 
   }
 
@@ -372,7 +400,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
           icon: FontAwesomeIcons.minusCircle,
           onSuccess: (data) async {
             _stockUpdateMessage(true);
-            refresh();
+            refresh(context);
           }
       );
 
@@ -413,7 +441,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
 
     _stockUpdateMessage(result);
 
-    refresh();
+    refresh(context);
   }
 
   Future <void> _countStockDialog() async {
@@ -445,7 +473,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
           icon: FontAwesomeIcons.clipboardCheck,
           onSuccess: (data) async {
             _stockUpdateMessage(true);
-            refresh();
+            refresh(context);
           }
       );
 
@@ -494,7 +522,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
       );
     }
 
-    refresh();
+    refresh(context);
   }
 
 
@@ -509,7 +537,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
 
     var result = await item.transferStock(context, locationId, quantity: quantity, notes: notes);
 
-    refresh();
+    refresh(context);
 
     if (result) {
       showSnackIcon(L10().stockItemTransferred, success: true);
@@ -549,7 +577,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
           icon: FontAwesomeIcons.dolly,
           onSuccess: (data) async {
             _stockUpdateMessage(true);
-            refresh();
+            refresh(context);
           }
       );
 
@@ -813,8 +841,8 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
                     context,
                     MaterialPageRoute(
                         builder: (context) => StockItemTestResultsWidget(item))
-                ).then((context) {
-                  refresh();
+                ).then((ctx) {
+                  refresh(context);
                 });
               }
           )
@@ -940,8 +968,8 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
           Navigator.push(
             context,
             MaterialPageRoute(builder: (context) => InvenTreeQRView(StockItemScanIntoLocationHandler(item)))
-          ).then((context) {
-            refresh();
+          ).then((ctx) {
+            refresh(context);
           });
         },
       )
@@ -971,7 +999,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
                     icon: Icons.qr_code,
                   );
 
-                  refresh();
+                  refresh(context);
                 }
               });
             });
@@ -1005,6 +1033,19 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
             _printLabel(context);
           },
         ),
+      );
+    }
+
+    // If the user has permission to delete this stock item
+    if (InvenTreeAPI().checkPermission("stock", "delete")) {
+      tiles.add(
+        ListTile(
+          title: Text("Delete Stock Item"),
+          leading: FaIcon(FontAwesomeIcons.trashAlt, color: COLOR_DANGER),
+          onTap: () {
+            _deleteItem(context);
+          },
+        )
       );
     }